Legal_Advisor_AI/requirements.txt
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
flask
langchain-community
request